requirements.txt
Django>=3.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary